livekit-agent/requirements.txt
livekit-agents[silero,turn-detector]~=1.2
livekit-plugins-deepgram~=1.0
livekit-plugins-cartesia~=1.0
livekit-plugins-openai~=1.0
python-dotenv>=1.0.0
